LayoutTests/inspector/dom/hideHighlight-expected.txt

Tests for the DOM.hideHighlight command.


== Running test suite: DOM.hideHighlight
-- Running test case: CheckEmptyHighlight
PASS: Should not be a highlight yet.

-- Running test case: HighlightRect
PASS: Should be one highlight rect.
Highlight Rect: {"top":0,"right":100,"bottom":100,"left":0,"width":100,"height":100}

-- Running test case: HideHighlight
PASS: Should be no highlight.

-- Running test case: HighlightQuad
PASS: Should be one highlight rect.
Highlight Rect: {"top":100,"right":150,"bottom":200,"left":50,"width":100,"height":100}

-- Running test case: HideHighlight
PASS: Should be no highlight.

-- Running test case: HighlightNode
PASS: Should be one highlighted node.
Highlighted Element Data: {"tagName":"p","idValue":"target","size":{"width":500,"height":100},"role":""}

-- Running test case: HideHighlight
PASS: Should be no highlight.

-- Running test case: HideHighlight
PASS: Should be no highlight.
